+++ src/menu/settings.ts
@@ -1,5 +1,6 @@
 import * as vscode from "vscode"
 import { Item } from "."
+import { loadGlobalConfig } from "../config"
 
 export const item: Item = {
   label: "$(settings) Settings",
@@ -8,6 +9,8 @@ export const item: Item = {
 }
 
 export function run() {
+  loadGlobalConfig()
+
   vscode.commands.executeCommand(
     "workbench.action.openSettings",
     "@ext:dervex.argon",